chore: move utils to infra-common
