--- infra-gen2/backends/auth/common/package.json
@@ -1,3 +1,0 @@
-{
-  "type": "module"
-}
