+++ infra-gen2/infra-common/src/auth-user-extensions/README.md
@@ -1,0 +1,6 @@
+# auth-user-extensions
+
+This directory contains extensions useful for managing authorized users, including:
+
+- creating users via App Sync
+- sending confirmation codes to App Sync instead of email/SMS
