+++ infra-gen2/infra-common/src/auth-user-extensions/add-create-user-lambda.ts
