+++ infra-gen2/infra-common/src/auth-user-extensions/add-custom-email-sender-lambda.ts
@@ -9,7 +9,7 @@ import { Runtime } from "aws-cdk-lib/aws-lambda";
 import { NodejsFunction } from "aws-cdk-lib/aws-lambda-nodejs";
 import path from "path";
 import url from "url";
-import { Mutable } from "../../common/types";
+import { Mutable } from "../types";
 
 const __filename = url.fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
