+++ infra-gen2/infra-common/src/auth-user-extensions/user-graphql-schema.graphql
