+++ infra-gen2/infra-common/src/expiration.ts
