+++ infra-gen2/infra-common/src/index.ts
@@ -1,3 +1,5 @@
-import { preSignUpTriggerHandler } from "./pre-sign-up";
+import { addAuthUserExtensions } from "./auth-user-extensions/add-auth-user-extensions";
+import { preSignUpTriggerHandler } from "./lambda-triggers/pre-sign-up";
 
-export { preSignUpTriggerHandler };
+export { addAuthUserExtensions, preSignUpTriggerHandler };
+
