+++ infra-gen2/infra-common/src/lambda-triggers/create-user.ts
