+++ infra-gen2/infra-common/src/lambda-triggers/custom-email-sender.ts
