+++ infra-gen2/infra-common/src/lambda-triggers/pre-sign-up.ts
