+++ infra-gen2/infra-common/src/types.ts
