+++ infra-gen2/package.json
@@ -13,6 +13,7 @@
   ],
   "dependencies": {
     "aws-amplify": "^6.4.3",
+    "@aws-crypto/client-node": "^4.0.0",
     "infra-common": "1.0.0"
   },
   "devDependencies": {
